Update tutorial files
