--- public/tutorial/data/emissions.csv
@@ -1,49 +1,0 @@
-,Country,Carbon dioxide,Greenhouse gases,Hydrofluorocarbons,Methane,Nitrogen trifluoride,Nitrous oxide,Perfluorocarbons,Sulphur hexafluoride,Unspecified mix of HFCs and PFCs
-0,Australia,388777.53,528631.65,11405.41,109934.74,,18060.12,291.48,162.36,
-1,Austria,66018.63,77532.35,1485.82,6499.26,11.96,3122.74,23.4,370.53,
-2,Belarus,60938.2,91988.2,208.96,17875.94,6.05,12938.02,10.54,10.49,
-3,Belgium,95668.08,110951.73,2397.3,7870.1,5.25,4733.21,182.0,95.79,
-4,Bulgaria,42281.88,53917.27,739.42,6560.6,,4312.08,0.01,23.27,
-5,Canada,537173.74,670428.3,11433.41,90509.76,0.62,30231.4,752.88,326.5,
-6,Croatia,17410.63,24446.42,1699.28,3887.16,,1439.72,,9.63,
-7,Cyprus,7029.36,8670.02,353.49,1097.4,,173.53,,16.24,
-8,Czech Republic,96665.23,118381.69,3711.4,13223.99,1.46,4683.99,30.95,64.68,
-9,Denmark,31306.35,45515.58,356.19,8644.7,,5192.77,0.01,15.57,
-10,Estonia,10419.92,12615.17,190.21,1152.94,,849.03,,3.07,
-11,European Union – 27 countries (from 01/02/2020),2814421.65,3468394.08,69863.8,401658.56,71.29,174381.19,1617.27,4837.8,1542.54
-12,Finland,37947.18,47799.49,837.6,4819.22,,4171.05,1.59,22.85,
-13,France,320908.3,420060.76,9984.62,62196.31,11.6,26031.4,390.18,538.34,
-14,Germany,678798.87,760358.01,8069.18,45687.75,13.72,24767.16,190.95,2663.74,166.64
-15,Greece,57556.33,77488.83,4675.45,11323.42,,3817.43,111.18,5.03,
-16,Hungary,48563.86,64217.84,1862.46,9206.95,,4485.77,1.76,97.04,
-17,Iceland,3510.05,4662.24,157.25,654.08,,248.94,88.95,2.97,
-18,Ireland,37547.28,62109.87,672.99,17649.7,12.07,6146.64,64.95,16.23,
-19,Italy,337229.88,417591.43,15387.77,47087.05,15.23,17193.34,395.32,257.5,25.33
-20,Japan,1062129.24,1168094.47,53561.04,27361.36,380.11,19459.73,3155.54,2047.45,
-21,Kazakhstan,255142.93,338123.36,2706.49,59921.35,,20340.56,9.65,2.37,
-22,Latvia,7240.5,10725.29,250.34,1889.57,,1332.78,,12.1,
-23,Liechtenstein,145.88,183.9,7.99,21.81,,8.16,0.0,0.05,
-24,Lithuania,13829.67,20251.51,512.85,3301.1,,2595.25,,12.63,
-25,Luxembourg,8429.71,9390.73,43.69,663.18,,244.23,,9.92,
-26,Malta,1606.73,2133.62,228.8,241.7,,55.96,0.01,0.41,
-27,Monaco,64.42,73.63,5.58,0.66,,2.85,0.0,0.12,
-28,Netherlands,139906.96,167153.25,1172.45,18715.26,,7155.25,79.45,123.87,
-29,New Zealand,34318.0,76824.59,1483.22,33019.35,,7937.6,50.69,15.73,
-30,Norway,40662.91,48901.96,628.42,5256.61,,2068.77,222.76,62.48,
-31,OECD - Europe,,4364518.0,,,,,,,
-32,OECD - Total,,14920010.0,,,,,,,
-33,OECD America,,8048043.0,,,,,,,
-34,OECD Asia Oceania,,2507448.5,,,,,,,
-35,Poland,331077.16,399438.49,4937.26,42704.71,,20616.67,10.08,92.6,
-36,Portugal,39938.33,56359.41,3158.99,10272.78,,2938.57,26.54,24.2,
-37,Romania,77190.58,115403.15,1908.24,25785.35,,10465.42,3.4,50.17,
-38,Russia,1711993.32,2156599.34,38619.93,314778.31,1.59,88400.86,1628.56,1176.77,
-39,Slovak Republic,35166.81,41226.49,672.37,3700.76,,1663.74,5.37,17.44,
-40,Slovenia,13059.97,16106.48,267.04,2085.12,,670.6,7.04,16.71,
-41,Spain,230269.46,288847.82,3612.74,41492.47,,11830.99,51.39,240.2,1350.56
-42,Sweden,38524.79,47816.7,840.83,4519.68,,3850.11,41.69,39.61,
-43,Switzerland,35787.34,45135.72,1241.48,5104.0,0.37,2845.14,28.37,129.03,
-44,Türkiye,452702.79,564389.75,7209.8,64020.23,,40306.09,6.79,144.05,
-45,Ukraine,210153.3,327258.77,1901.02,71536.71,,43618.8,,48.94,
-46,United Kingdom,348409.72,429489.47,10819.71,51738.66,0.34,17906.35,192.88,421.82,
-47,United States,5032212.82,6340228.29,158774.99,727444.74,636.46,393279.02,3541.3,8018.58,16320.39
